Initialize int_sync and irq_flags in non DT use case.
Both fields need to be populated in the non device tree use case to make
the touch controller work.

When int_sync is not set to true,
the initialization sequence is never finished by the host cpu
and the controller remains in limbo.

If the irq_flags are unassigned, falling edge triggered is assumed but
the controller does rising edge triggered interrupts by default.